better test setup
